requirements.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1
flask-cors==4.0.0
